requirements.txt
# === Core Dependencies ===
streamlit==1.35.0
openai==1.30.1
python-dotenv==1.0.1

# === Optional: For JSON handling, better structure, async utils ===
pydantic==2.7.1
aiofiles==23.2.1

# === NLP (if adding advanced tone detection later) ===
nltk==3.8.1

# === Whisper audio (future feature) ===
# openai-whisper==20231117

# === Data/Plotting for Premium Dashboard (Block 5) ===
pandas==2.2.2
matplotlib==3.9.0
altair==5.3.0
